Adds and refactors code
