+++ test/Microsoft.Health.Fhir.Shared.Tests.E2E/Rest/Audit/AuditTests.cs
@@ -21,9 +21,11 @@ using Task = System.Threading.Tasks.Task;
 
 namespace Microsoft.Health.Fhir.Tests.E2E.Rest.Audit
 {
-    [Trait(Traits.Category, Categories.Batch)]
+    /// <summary>
+    /// Provides Audit specific tests.
+    /// </summary
     [HttpIntegrationFixtureArgumentSets(DataStore.CosmosDb, Format.Json)]
-    public class AuditTests : IClassFixture<AuditTestFixture>
+    public partial class AuditTests : IClassFixture<AuditTestFixture>
     {
         private const string RequestIdHeaderName = "X-Request-Id";
         private const string CustomAuditHeaderPrefix = "X-MS-AZUREFHIR-AUDIT-";
@@ -363,28 +365,6 @@ namespace Microsoft.Health.Fhir.Tests.E2E.Rest.Audit
                 expectedAppId: TestApplications.NativeClient.ClientId);
         }
 
-        [Fact]
-        [Trait(Traits.Priority, Priority.One)]
-        public async Task GivenABatch_WhenPost_ThenAuditLogEntriesShouldBeCreated()
-        {
-            if (!_fixture.IsUsingInProcTestServer)
-            {
-                // This test only works with the in-proc server with customized middleware pipeline
-                return;
-            }
-
-            var expectedActions = new List<string> { "batch", "delete", "delete", "create", "create", "create", "create", "update", "update", "update", "update", "update", "update", "search-type", "search-type", "read", "read", "batch" };
-            var expectedPathSegments = new List<string> { string.Empty, "Patient/234", "Patient/234", "Patient", "Patient", "Patient", "Patient", "Patient/123", "Patient/123", "Patient?identifier=234234", "Patient?identifier=234234", "Patient/123a", "Patient/123a", "Patient?name=peter", "Patient?name=peter", "Patient/12334", "Patient/12334", string.Empty };
-
-            // Only status codes at even, first and last positions are validated, since these are executed audit entries
-            var expectedStatusCodes = new List<HttpStatusCode> { HttpStatusCode.OK, HttpStatusCode.OK, HttpStatusCode.NoContent, HttpStatusCode.OK, HttpStatusCode.Created, HttpStatusCode.NoContent, HttpStatusCode.Created, HttpStatusCode.OK, HttpStatusCode.OK, HttpStatusCode.OK, HttpStatusCode.Created, HttpStatusCode.OK, HttpStatusCode.NotFound, HttpStatusCode.OK, HttpStatusCode.OK, HttpStatusCode.NotFound, HttpStatusCode.NotFound, HttpStatusCode.OK };
-            await ExecuteAndValidateBatch(
-               () => _client.PostBundleAsync(Samples.GetDefaultBatch().ToPoco()),
-               expectedActions,
-               expectedPathSegments,
-               expectedStatusCodes);
-        }
-
         private async Task ExecuteAndValidate<T>(Func<Task<FhirResponse<T>>> action, string expectedAction, ResourceType expectedResourceType, Func<T, string> expectedPathGenerator, HttpStatusCode expectedStatusCode)
             where T : Resource
         {
@@ -448,13 +428,11 @@ namespace Microsoft.Health.Fhir.Tests.E2E.Rest.Audit
                     {
                         expectedResourceType = ResourceType.Patient;
                     }
-
-                    if (auditList[i].StatusCode.Equals(HttpStatusCode.NotFound))
+                    else if (auditList[i].StatusCode.Equals(HttpStatusCode.NotFound) || auditList[i].StatusCode.Equals(HttpStatusCode.PreconditionFailed))
                     {
                         expectedResourceType = ResourceType.OperationOutcome;
                     }
-
-                    if (auditList[i].StatusCode.Equals(HttpStatusCode.OK) && (expectedActions[i].Equals("search-type") || expectedActions[i].Equals("batch")))
+                    else if (auditList[i].StatusCode.Equals(HttpStatusCode.OK) && (expectedActions[i].Equals("search-type") || expectedActions[i].Equals("batch")))
                     {
                         expectedResourceType = ResourceType.Bundle;
                     }
